Base initialization
This is an alternative example for ofxKinect which uses the thresholds
to generate a mask that defines which points in the kinect point cloud
are used. It also uses an alternative stepping value method, and uses
ofxDelaunay to turn this points into triangles. It relies on ofxGui to
change some values instead of keyboard commands.